Delete Util en model
