+++ SoftModel/src/main/java/com/mycompany/softmodel/CitaDTO.java
@@ -4,8 +4,6 @@
  */
 package com.mycompany.softmodel;
 
-import com.mycompany.softmodel.Util.Estado;
-
 /**
  *
  * @author Mcerv
