+++ SoftModel/src/main/java/com/mycompany/softmodel/Rol.java
@@ -2,12 +2,12 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.mycompany.softmodel.Util;
+package com.mycompany.softmodel;
 
 /**
  *
  * @author Mcerv
  */
-public enum Turno {
-    MANHANA, TARDE
+public enum Rol {
+    PACIENTE, MEDICO, ADMINISTRADOR
 }
